% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dobble.R
\name{cards_to_A4}
\alias{cards_to_A4}
\title{Assembling cards on DINA4 sheets}
\usage{
cards_to_A4(cardfolder, outfolder)
}
\arguments{
\item{cardfolder}{is a folder containing the 57 individual cards}

\item{outfolder}{is the folder in which the final DIN A4 pdf sheets are saved}
}
\description{
Assembling cards on DINA4 sheets
}
