% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dobble.R
\name{dobble_template}
\alias{dobble_template}
\title{dobble_template}
\usage{
dobble_template()
}
\value{
vector
}
\description{
Creates a list of dobble cards using numbers instead of images
}
\examples{
dobble_template()
}
