% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_to_cards.R
\name{folder_to_cards}
\alias{folder_to_cards}
\title{Folder to cards}
\usage{
folder_to_cards(infolder, outfolder, output = "SINGLE")
}
\arguments{
\item{output}{specifies whether each individual card should be saved or whether DINA4 sheets with six cards on each should be saved}
}
\value{
a set of dobble cards saved in the outfolder
}
\description{
Folder to cards
}
