% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_all_cards.R
\name{generate_all_cards}
\alias{generate_all_cards}
\title{Generate all cards}
\usage{
generate_all_cards(infolder, saving = FALSE, outfolder)
}
\arguments{
\item{outfolder}{is the path of a folder in which the 57 created cards are saved}

\item{x}{is a list of dobble card templates}
}
\value{
a set of dobble cards saved in the outfolder
}
\description{
This function generates all dobble cards with the help of the generate one card function
}
