% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_one_card.R
\name{generate_one_card}
\alias{generate_one_card}
\title{Generate one card with ggplot}
\usage{
generate_one_card(x, infolder, file_list, positions)
}
\arguments{
\item{x}{a vector of card numbers}

\item{infolder}{the folder to read the files from}

\item{positions}{a dataframe of positions for the images on the graph}
}
\value{
one dobble card as ggplot object
}
\description{
This function generates one dobble card with the help of the circlize package.
}
