% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_load_transform.R
\name{image_load_transform}
\alias{image_load_transform}
\title{Load and apply some transformations to the images}
\usage{
image_load_transform(x, infolder, file_list)
}
\arguments{
\item{x}{a vector of image numbers}

\item{infolder}{the folder to read the files from}
}
\value{
a picture
}
\description{
This function generates one dobble card with the help of the circlize package.
}
