% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picture_size_generate.R
\name{picture_size_generate}
\alias{picture_size_generate}
\title{Generate a list of picture sizes}
\usage{
picture_size_generate(x)
}
\arguments{
\item{x}{an image}
}
\value{
a vector of imagesizes
}
\description{
Generate a list of picture sizes
}
