% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_all_cards.R
\name{generate_all_cards}
\alias{generate_all_cards}
\title{Generate all cards}
\usage{
generate_all_cards(image_list)
}
\value{
a set of dobble cards saved in the outfolder
}
\description{
This function generates all dobble cards with the help of the generate one card function
}
